chore: refactor performance test and add script for testing performance with debugger attached
